harvest::routes::insert_cast_vote

Function insert_cast_vote

pub async fn insert_cast_vote(
    body: Json<InsertCastVoteInput>,
    claims: JwtClaims,
    user_info: UserLocation,
) -> Result<Json<InsertCastVoteOutput>, Custom<Json<ErrorResponse>>>

API endpoint for inserting votes. POST coming from the frontend->Hasura->Harvest->Here.

It tries to insert the vote into the database and returns the Json result in case of success or logs the information of the error (coming from a synchronous windmill library function, try_insert_cast_vote) before returning the error.